+++ src/main/java/projlab/IStep.java
@@ -2,9 +2,9 @@ package projlab;
 
 //TODO: Add Javadoc
 public interface IStep {
-    boolean onStepped(Player player, Direction direction, int power);
+    boolean onStepped(Player player, Direction direction, double power);
 
-    boolean onStepped(Box box, Direction direction, int power);
+    boolean onStepped(Box box, Direction direction, double power);
 
     Field offStepped(Player player, Direction direction);
